# openapi-auction.yaml
openapi: 3.1.0
info:
  title: Auction Service API
  version: 1.0.0
  description: |
    English-style (ascending, open outcry) auction API designed for a Go (transactional REST)
    + Node.js (WebSocket gateway) + MySQL stack. Includes anti-sniping (soft-close),
    blacklist enforcement, idempotent writes, and event stream endpoints.

    ## Conventions
    - **Auth**: JWT via `Authorization: Bearer <token>` or cookie `session` (see `cookieAuth`).
    - **Time**: ISO 8601 UTC (e.g., `2025-09-15T02:00:00Z`).
    - **Idempotency**: All mutating endpoints that can be retried must include `Idempotency-Key` header.
    - **Anti-sniping**: Within `soft_close_window_secs` of `end_time`, a valid bid extends the auction by `soft_close_extend_secs`.
    - **Anonymized bidders**: If `anonymize_bidders=true`, clients should mask bidder identity in UI (server still returns IDs unless `mask_bidders=true` query is used).
servers:
  - url: https://api.example.com
    description: Production
  - url: https://staging.api.example.com
    description: Staging
  - url: http://localhost:8080
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    AuctionId:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64, minimum: 1 }
      description: Auction ID
...
    MaskBidders:
      name: mask_bidders
      in: query
      schema: { type: boolean, default: false }
      description: If true, server masks bidder identity in responses where applicable
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - 400_BAD_REQUEST
                - 400_UNDER_MIN_INCREMENT
                - 400_INVALID_STATE
                - 400_INVALID_TIME_RANGE
                - 401_UNAUTHORIZED
                - 403_BLACKLISTED
                - 403_FORBIDDEN
                - 404_NOT_FOUND
                - 409_CONFLICT
                - 409_ACTIVE_AUCTION_EXISTS
                - 409_OUTBID_RACE
                - 409_AUCTION_NOT_RUNNING
                - 409_BUY_IT_NOW_ONLY
                - 422_UNPROCESSABLE_ENTITY
                - 500_INTERNAL_ERROR
            message:
              type: string
            meta:
              type: object
              additionalProperties: true
...
    AuctionType:
      type: string
      enum: [ENGLISH_OPEN]
      default: ENGLISH_OPEN
    AuctionStatus:
      type: string
      enum: [DRAFT, SCHEDULED, RUNNING, ENDED, CANCELLED]
...
    AuctionDetail:
      type: object
      required:
        - id
        - listing_id
        - seller_id
        - status
        - start_time
        - end_time
        - min_increment
        - currency
        - soft_close_window_secs
        - soft_close_extend_secs
        - anonymize_bidders
      properties:
        id: { type: integer, format: int64 }
        listing_id: { type: integer, format: int64 }
        seller_id: { type: integer, format: int64 }
        type: { $ref: '#/components/schemas/AuctionType' }
        status: { $ref: '#/components/schemas/AuctionStatus' }
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        reserve_price: { type: number, format: double, nullable: true }
        min_increment: { type: number, format: double }
        buy_it_now: { type: number, format: double, nullable: true }
        soft_close_window_secs: { type: integer }
        soft_close_extend_secs: { type: integer }
        anonymize_bidders: { type: boolean }
        current_price: { type: number, format: double, nullable: true }
        highest_bidder_id: { type: integer, format: int64, nullable: true }
        reserve_met: { type: boolean, default: false }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    AuctionUpdate:
      type: object
      description: Patch fields allowed while auction has not started (DRAFT/SCHEDULED)
      properties:
        start_time: { type: string, format: date-time }
        end_time: { type: string, format: date-time }
        reserve_price: { type: number, format: double, nullable: true }
        min_increment: { type: number, format: double }
        buy_it_now: { type: number, format: double, nullable: true }
        soft_close_window_secs: { type: integer, minimum: 0 }
        soft_close_extend_secs: { type: integer, minimum: 0 }
        anonymize_bidders: { type: boolean }
...
paths:
...
  /api/v1/auctions/{id}:
    get:
      tags: [Auctions]
      summary: Get auction detail
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AuctionId'
        - $ref: '#/components/parameters/MaskBidders'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuctionDetail' }
        '404':
          description: Not Found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    patch:
      tags: [Auctions]
      summary: Update editable fields (before start)
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/AuctionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuctionUpdate' }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/AuctionDetail' } } }
        '400': { description: Bad Request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409': { description: Conflict, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
security:
  - bearerAuth: []
  - cookieAuth: []
...